wip(slice01): S6 verification — split navigation codec under size threshold; land identity/navigation docs
